fix: load page occours logout
