+++ src/pages/login/index.tsx
@@ -1,9 +1,14 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { HeaderLoginAndRegister } from "../../components/macro-components/header-login-and-register";
 import { UsernameModal } from "./username-modal";
 import { UserPasswordModal } from "./userpassword-modal";
 import { FooterLoginAndRegister } from "../../components/macro-components/footer-login-and-register";
 import { ForgotPassword } from "./forgot-password-modal";
+import { useDispatch } from "react-redux";
+import { useNavigate } from "react-router";
+import { api } from "../../lib/axios";
+import { setLogin } from "../../features/login-slice";
+import { setUser } from "../../features/user-slice";
 export function Login() {
 
     const [filledUserName, setFilledUserName] = useState(false);
@@ -12,8 +17,40 @@ export function Login() {
 
     const [ username, setUsername ] = useState('');
     const [ password, setPassword ] = useState('');
-
     const [ textAlert, setTextAlert ] = useState('');
+
+    const dispatch = useDispatch();
+    const navigate = useNavigate();
+
+        
+    async function loadStorage() {
+        const token = localStorage.getItem('token');
+        
+        if (token) {
+            await api.get(`/v3/users/current`, {
+                headers: {
+                    'Authorization': `Bearer ${token}`
+                }
+            })
+            .then((json) => {
+                api.defaults.headers['Authorization'] = `Bearer ${token}`;
+                dispatch(setUser(json.data));
+                dispatch(setLogin(true));
+                navigate('/home');
+                
+            })
+            .catch((err) => {
+                dispatch(setLogin(false));
+                console.log(err);
+            });
+        }
+    }
+    
+    useEffect(() => {
+        loadStorage();
+    }, []);
+
+
 
 
     function openForgotPassword() {
